+++ worldeditadditions_commands/commands/nodes/init.lua
@@ -9,3 +9,4 @@
 local we_cmdpath = worldeditadditions_commands.modpath .. "/commands/nodes/"
 
 dofile(we_cmdpath .. "ngroups.lua")
+dofile(we_cmdpath .. "ndef.lua")